requirements.txt
librosa
soundfile
scikit-learn>=1.3.2
funasr>=1.1.2
moviepy
numpy==1.26.4
gradio
modelscope
torch>=1.13
torchaudio
openai
g4f
dashscope
curl_cffi

moviepy
Pillow
python-dotenv